+++ CHANGES.md
@@ -4,6 +4,9 @@ Changes
 1.3.5.4 (unreleased)
 --------------------
 
+- The resource reference from the `MessageHeader` to the so called focal
+  resource of the bundle needs to be versioned.
+
 - Have the `InvalidValue` exception report a tiny bit more information.
 
 - Improvements for messages subcommand.